+++ WebContent/META-INF/SQL/product.sql
@@ -4,8 +4,8 @@ create table product_table(
 p_num number,
 p_name varchar2(20) not null,
 p_contents varchar2(1000) not null,
-p_price int(10) not null,
-p_quantity int(10) not null,
+p_price number(10) not null,
+p_quantity number(10) not null,
 p_color varchar2(10) not null,
 primary key(p_num)
 );
